[dataset] simplify saving/reading of TLVs in/from secure storage
This commit simplifies the handling of Dataset TLVs that need to be
securely stored when the `PLATFORM_KEY_REFERENCES_ENABLE` feature is
enabled. Two new methods are added to the `Dataset` class:

- `SaveTlvInSecureStorageAndClearValue()` which  saves the value of a
  given TLV type in secure storage and clears the TLV value by
  setting all value bytes to zero.

- `ReadTlvFromSecureStorage()` which reads the TLV value back from
  secure storage and updates it in the `Dataset`.

These methods are used by `DatasetLocal` to manage the set of TLVs
that need to be securely stored defined by a constant array of
`SecurelyStoredTlv` entries. Each entry provides the TLV type
along with the associated `Crypto::Storage::KeyRef` to use for
active or pending Dataset.