[mesh-forwarder] drop msg if frag tx delay exceeds reassembly timeout
This commit introduces a new mechanism in `MeshForwarder` to drop a
`Message` if the transmission delay between its fragments exceeds the
reassembly timeout (2 seconds). This improves performance by
preventing the transmission of fragments that are unlikely to be
reassembled successfully at the receiver.

This commit also adds a new `MessageAction` to log such message drops
for better visibility and debugging.

Finally, this commit removes the previous config option
`OPENTHREAD_CONFIG_6LOWPAN_REASSEMBLY_TIMEOUT`, fixing the timeout to
the value of 2 seconds.