[posix] enable building `infra_if.cpp` on macOS (#9891)
This change makes it possible to build the `posix/platform/infra_if`
module on macOS (and other non-Linux systems). It adds guard checks
`#ifdef __linux__` around the use of `mNetLinkSocket`(used for
network interface state change detection). This allows compilation to
proceed on macOS, though equivalent functionality for detecting netif
change is not yet implemented.

This commit also adds a new `toranj-macos` workflow job to validate
that the code builds successfully with the `posix` platform.