[version] introduce `OT_THREAD_VERSION_1_4`
Update constants and build switch options to reflect the new version number "1.4" (previously "1.3.1").

Keeping the legacy "1.3.1" and OT_THREAD_VERSION_1_3_1 types for backward compatibility with projects that might be using those version names.

Note that this commit does not change the default config value of `OPENTHREAD_CONFIG_THREAD_VERSION` which is the default version to use if it is not explicitly specified by build switches or in a project config file.