[mle] retain direct child cache entries on Addr Solicit Response TX
Previously, when a child was promoted to a router role, the
leader upon sending "Address Solicit Response" would clear cache entries
associated with the child's old RLOC16. This commit modifies the code
to retain these cache entries when the sender is a direct child
of the leader. By retaining the cache entries, they can be associated
with the promoted router's new RLOC16 when the Link Advertisement is
received.