[cli] move `Process{Get/Set/Enable/Disable}` methods to Utils class (#9951)
This commit moves the `Process{Get/Set/Enable/Disable}` methods to the
Output class and rename the `Output` class to `Utils`.  The purpose is
to reduce code size in the Interpreter class to make the class focus
more on the command handling and easier for further refactoring.

The commit also removes the alias typedef of `Arg` in each cli module
because they are unnecessary. Having one definition in the parent
class is enough.