[cli] `netdata show` to support filtering by RLOC16
This commit updates the CLI `netdata show` command to allow an
optional RLOC16 as input. Providing an RLOC16 filters the output to
display prefix, route, and service entries associated specifically
with the specified RLOC16. This is helpful for checking entries added
by a specific border router.

`test-011-network-data-timeout.py` is updated to validate the new CLI
functionality.