[ip6] allow delivering RLOC/ALOC traffic to host
Filtering out traffic to/from RLOC/ALOC was added by #423 for #419.
The intention was to filter out Thread's control traffic.

Now we already added filtering out traffic by checking whether it
is destined to a service hosted in OpenThread stack, so filtering
out by RLOC/ALOC is not necessary anymore.

This commit removes this filter, as a result, the core lib gets
rid of the `PLATFORM_NETIF` flag.