[cli] update README files to clarify a Commissioner is used for changing datasets (#10026)
Based on discussion in #9901 , this clarifies in the documentation
that the CLI `dataset` commands can only be used in specific
situations (like testing), and that these commands allow setting
invalid (combinations of) parameters also. It points to the
Commissioner as the correct way of doing dataset updates in
production.

It also adds "Overview" sections that were missing in two README
files.