[routing-manager] add support for to use a prefix received from DHCPv6 PD (#10000)
This commit adds support for adding a DHCPv6 delegated prefix to the
Routing Manager via a new Open Thread API. Previously this feature was
only supported using a platform generated RA on a Posix based BR. For
platforms that can get the prefix directly from a DHCPv6 PD client API
there is no point in creating a RA for setting the prefix. It's much
easier to use a direct call to configure it.

This new API uses an otBorderRoutingPrefixTableEntry structure as an
input parameter and evaluates the prefix based on the same rules as
one coming from a platform RA.  The lifetime of the prefix can be
updated by calling the function again with new time values.