[routing-manager] ensure correct handling of shorter PD prefix
This commit corrects the order of calls to prefix methods `Tidy()` and
`SetLength(kOmrPrefixLength)`. This was unintentionally reversed in
PR #10000.

PD prefixes can be shorter than `kOmrPrefixLength` (as allowed by
`IsValidPdPrefix()`). To use them as OMR prefixes, `Tidy()` is called
first to clear any extra bits before potentially extending the prefix
to `kOmrPrefixLength`.