[spinel] only request crash logs on RCP recovery if capable (#10213)
During RCP recovery, the Host does not check if the RCP supports
logging a crash dump. This causes RCP recovery to fail when the RCP is
built with `OPENTHREAD_CONFIG_PLATFORM_LOG_CRASH_DUMP_ENABLE =
0`. This bug was introduced in #10061

This commit will make the Host only request crash logs from the RCP if
the RCP supports it.