[routing-manager] rename `DiscoveredPrefixTable` to `RxRaTracker`
This commit renames the `DiscoveredPrefixTable` to `RxRaTracker`. The
new name is shorter and emphasizes that this class is responsible for
processing received RA messages and tracks useful information about
the RA (routers and their advertised flags and prefixes).

It also renames `RaInfo` to `TxRaInfo` to clarify its role for
tracking information about emitted RAs.