[mdns] ensure callback is invoked when registering host with no address
This commit updates the native mDNS implementation to ensure the
callback is invoked when a host is registered without any address
(effectively unregistering the host and removing any previously
registered addresses for the host-name).

The implementation ensures the callback is invoked after returning
from the `RegisterHost()` call (invoked from a posted tasklet), as
required by the mDNS API definitions.

Additionally, the `test_mdns` unit test is updated to cover
registering a host with no address for the first time or updating a
previous registration.