[channel-manager] require CSL to enabled with `CSL_CHANNEL_SELECT_ENABLE`
This commit requires `MAC_CSL_RECEIVER_ENABLE` to be enabled when
`CHANNEL_MANAGER_CSL_CHANNEL_SELECT_ENABLE` is also enabled. This is
checked and enforced in the `channel_manager.hpp`, emitting an
`#error` if it fails. This simplifies the `#if` checks related to
this in the `ChannelManager` class and CLI module.