[routing-manager] update stale time constant to 10 minute
This commit renames the entry stale time constant to `kStaleTime` and
sets it to 10 minutes (from 30 minutes). This change aligns the
`RoutingManager` with the recommended value in the latest stub router
RFC draft (`STALE_RA_TIME`).

The unit test `test_routing_manager` is also updated to reflect this
new constant.